修复编译警告：消除 -Wcast-qual/-Wswitch-enum/-Wcovered-switch-default 等警告，提升警告等级至 -W3，添加第三方库抑制选项，实现零警告构建

+++ stm32f407/src/i2c_drv.c
@@ -29,7 +29,7 @@ int i2c_drv_write_reg(uint8_t dev_addr, uint8_t reg, const uint8_t *data, uint16
         (uint16_t)(dev_addr << 1),
         (uint16_t)reg,
         I2C_MEMADD_SIZE_8BIT,
-        (uint8_t *)data,
+        (uint8_t *)(uintptr_t)data,
         len,
         I2C_TIMEOUT_DEFAULT
     );
